restolingo/requirements.txt
Django==1.10.3
django-autoslug
